requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.0
django-cors-headers==4.3.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
